# Conflicts:
#	CesiProgSys/ViewCli/ViewCliFr.cs

+++ CesiProgSys/ToolsBox/Info.cs
@@ -13,7 +13,7 @@ public class Info
     public string CurrentSource { get; set; }
     public string CurrentDest { get; set; }
     public int TotalFilesToCopy { get; set; }
-    public int TotalFilesSize { get; set; }
+    public long TotalFilesSize { get; set; }
     public int NbFilesLeftToDo { get; set; }
     public State state { get; set; }
     public float progression { get; set; }
